requirements.txt
Django==3.2.12
Pillow==9.5.0

